Refactor: Integrate logout functionality directly into MainNav component and remove LogoutButton
